% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testroxy-package.R
\docType{package}
\name{testroxy-package}
\alias{testroxy}
\alias{testroxy-package}
\title{testroxy: A Mock Package}
\description{
This os a full text to check the implementation of autolinking urls by roxygen2. The examples are taken from CRAN packages. Example on urls: See a non secured link on 'IRISSeismic', \url{http://service.iris.edu/}, regular link on 'abjutils' \url{https://abj.org.br/}, a link with masks on 'MEDITS' \url{https://www.sibm.it/SITO\%20MEDITS/principaleprogramme.htm} and the last with different masks from 'ggsolvencyii' \url{https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=OJ:L:2015:012:TOC}. Let's talk about dois: a regular doi from 'abclass' \doi{10.1093/biomet/asu017}, with caps from 'zeitgebr' \doi{10.1086/160554} and with masks from 'raytracing' \doi{10.1175/1520-0469(1996)053<2365:PORWON>2.0.CO;2}. By last, arXiv: One from 'abess' \href{https://arxiv.org/abs/1702.04690}{arXiv:1702.04690}, different casing by 'DLL' \href{https://arxiv.org/abs/1907.12732}{arXiv:1907.12732}, strange on from 'grove' \href{https://arxiv.org/abs/1602.03990v2}{arXiv:1602.03990v2 [stat.ME]} and old-style from 'QGameTheory' \href{https://arxiv.org/abs/quant-ph/0208069}{arXiv:quant-ph/0208069}.
}
\author{
\strong{Maintainer}: John Doe \email{john.doe@example.com}

}
\keyword{internal}
